add a 'logger' local to all templates
This exposes the engine's logger to the templates as a local so
you can log from inside your template renders.  Previously there
was no way to log output from the templates themselves.

By default, `logger` is the name of the local but this can be
overridden with the `'logger_local'` option.