update render api to take an option content block
This is the standard Deas expects.  In this case it is ignored.
This means rendering with layouts is not supported.

This is really just about having this engines implementation match
the base Deas template engine spec.  The render method can take
a content block, whether the engine does anything with it or not.