update to latest Deas render API; test cleanups
This brings in the latest Deas and conforms to its render API.  This
means taking content blocks on the partial method and removing the
capture partial method.

This also cleans up the tests by organizing the the tests that actually
render as system tests.  This also better commonized the setup logic
for those tests.